dags/data-transformation-dbt/requirements.txt
dbt-core==1.6.5
dbt-postgres==1.6.5
python-dotenv==1.0.0
PyYAML==6.0.1
SQLAlchemy==2.0.9
pandas==2.1.1